+++ CHANGELOG.md
@@ -8,3 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.1.0] - 2025-09-25
+
+## [1.0.0] - 2026-01-03
+
+### Added
+
+- Binance.US Spot exchange
+- Documentation for exchanges
+- `Exchange::cancel_order` now receive `Order` as argument
+- Use new
+  Binance [Algo API](https://developers.binance.com/docs/derivatives/usds-margined-futures/trade/rest-api/New-Algo-Order)
+  for stop orders
+- Listen key refresh for Binance Futures
+- Config folder changed `$HOME/.scalper/config` -> `$HOME/.scalper-rs/config`